+++ Gw2Sharp/Connection.cs
@@ -20,6 +20,7 @@ namespace Gw2Sharp
         private ICacheMethod cacheMethod;
         private ICacheMethod renderCacheMethod;
         private string accessToken;
+        private string apiBaseUrl = "https://api.guildwars2.com/";
         private readonly ObservableCollection<IWebApiMiddleware> middleware = new ObservableCollection<IWebApiMiddleware>();
 
 
@@ -130,6 +131,16 @@ namespace Gw2Sharp
 
         /// <inheritdoc />
         public string UserAgent { get; }
+
+        /// <inheritdoc />
+        public string ApiBaseUrl
+        {
+            get => this.apiBaseUrl;
+            set => this.apiBaseUrl = value ?? throw new ArgumentNullException(nameof(value), "ApiBaseUrl cannot be null");
+        }
+
+        /// <inheritdoc />
+        public string? RenderBaseUrl { get; set; }
 
         /// <inheritdoc />
         public TimeSpan RequestTimeout